A mildly simple approach at introducing the 'combined diff format'
This adds in a new diff format parser based on the combined diff format:

https://git-scm.com/docs/git-diff

Git generates this when you perform a 'git diff --cc' or 'git diff -c' invocation.  It shows the differences when the file had more than 1 parent.  Without this fix, such differences are ignored.

This is to help support fixing [issue 1028](https://github.com/gitleaks/gitleaks/issues/1028) in GitLeaks.  GitLeaks will also need to add the '--cc' argument, as mentioned by that issue.

There are a few limitations with this.  A big one is in the file header termination detection.  For now, it only scans for the 1-parent style ('@@ -') and the 2-parent style ('@@@ -').  The combined diff format allows for having many more than just 2 parents, so this will skip those cases.  I've updated the README to note about this limitation.

This includes some unit test coverage.  It's not exhaustive, but it does lend some credibility to the code.