Modifications based on real-world data
I found some situations where the 'diff --cc' can return a line count number of 18446744073709551615, which is a uint64 value of a signed int64 -1.  The code now adjusts for this strange result.

Additionally, a scenario came up where the line counting algorithm was off by 1, which may be related to the -1 issue above.  This was captured in the 'strangeCount' test.

Both of these scenarios point to a bug in the git diff algorithm.  It's due to this that the final line count check can't be used.