Handle empty commit headers
In some cases, especially for some merges, a commit can contain zero files.  This can throw off the header parser by interpreting the following headers as the body text of the first commit.  Down-stream, this causes the GitLeaks tool to identify the wrong commit ID with the leaked data.

This patch currently fails the empty-commit check.  A minor issue with the expected data eludes me at the moment.